+++ database/migrations/2025_09_18_212705_create_purchase_orders_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->date('tanggal');
             $table->date('tanggal_perkiraan_datang')->nullable();
             $table->string('status',20)->default('draft');
-            $table->decimal('total')->default(0);   
+            $table->unsignedBigInteger('total')->default(0);   
             $table->integer('approval_by')->nullable();   
             $table->string('approval_status',15)->default('pending');   
             $table->dateTime('approved_at')->nullable();
